refactor(convert): SoA 全链路 + SPZ v1-v4 读取 + SOG v3 + 压缩 PLY + round-trip 套件
- C-01: 解析直出 GaussianCloudSoA (ply/splat/spz/sog)，写入器消费 SoA
- C-02/TD-32: PLY 流式分块解析 (PLY_STREAM_CHUNK_ROWS)
- C-03: spz-reader.ts 支持 SPZ v1-v4 (NGSP/zstd 注入点)，11 用例
- C-04/TD-19: SOG v3 SH overlay 读写闭环 (SOG_MAGIC_V3 + readShOverlaySoA)
- C-05: compressed-ply-writer.ts SuperSplat 兼容压缩 PLY 输出
- C-07: CLI --max-splats 转换期预裁剪 + 质量回执; --sog-version 3
- C-09: round-trip 回归套件 (17 用例, 量化误差阈值断言)
- C-10: batch CLI 多格式输入 + manifest.json
- TD-15: contributionCutoff 改 quickselectIndices (O(N) 索引分区)

+++ packages/convert/src/spz-writer.test.ts
@@ -225,7 +225,11 @@ describe('writeSpz — Body 数据验证', () => {
     // Alpha 在 positions 之后: offset = numSplats * 9
     const alphaOffset = 5 * 9;
     for (let i = 0; i < 5; i++) {
-      const expectedAlpha = Math.max(0, Math.min(255, Math.round(cloud.splats[i].opacity * 255)));
+      // ★ C-01: 写入路径消费 SoA (f32 列式), 故期望值按 f32 精度计算
+      const expectedAlpha = Math.max(
+        0,
+        Math.min(255, Math.round(Math.fround(cloud.splats[i].opacity) * 255)),
+      );
       expect(body[alphaOffset + i]).toBe(expectedAlpha);
     }
   });
